Delete NavigateToURL.java
This file validates the  “illegal” strings are being stripped out of the “profile essay” fields (proper URLs and email addresses). Strings are matched before and after , and are validated if  @anything or www.anything.com is not displayed in the essay fields. Also it validates any editing changes that are further reflected in view profile pages.